Update applying-custom-policies.adoc
Removed the description of type in the policy YAML. Reason: type is optional, and supported options are system or custom. However system is used internally; and users are not able to upload custom policies with type: system. This document describes the creation of custom policies, so it will confusing for users.
I also removed the type: system from the example... doing a copy & paste of that example will not work.