Update information about installing MUnit on Studio
Due that MUnit comes by default with Studio since 5.4.0, there is no need to explain how to install it at first. Just in case that they are using a previous version of Studio, we are leaving it at the bottom of the page just to inform them how to do it. This will need to be replicated on version 1.1.1 and 1.0.0