Removed error handling from TOC
none of the things in that doc can be done in Design Center. They’re
all exclusive for Studio. There are explanations that still do make
sense in Design Center, but that can’t be acted on in any way in the
present release.

I hence removed the link from the TOC and any other incoming links to
the page