"Avoiding Public Discoverability" section updated
"Avoiding Public Discoverability for Applications on CloudHub" section is a bit confusing. Readers could easily conclude that they need dedicated load balancers in order to prevent their applications deployed in VPC from being publicly accessible. A whitelist in a dedicated load balancer is something different and should not be confused with public accessibility to the applications in VPC.

Therefore, I added a sentence emphasizing this.