make components more explicit
We were discussing with UX the possibility of not mentioning "cards" at all, as people find it confusing that these same things are referred to in various different names. For the time being we'll still use the word "card", but only when strictly referring to the UI element.

I made the relationship between cards and components explicit here, and then referred to the components instead.