Update to-create-an-asset.adoc
Made minor changes. Also the diagram is a little off using the terms. There is no OAS, RAML or WSDL API - those are specification formats. APIs are REST, SOAP and HTTP. I would change the APIs to specs that point to Exchange and remove APIM from the picture.

If you want to keep the APIM there, only keep the top container and next to the arrow from Exchange to APIM say "Automatic indexing of API endpoints"