Update configuring-properties.adoc
The reason for adding this change is because the HTTP URL's are allowed as a value in the location attribute, since the backing spring library supports URLs & URI's to pass. 

org.springframework.core.io.Resources class has a method getURL() which is called from org.springframework.core.io.support.PropertiesLoaderSupport. This PropertiesLoaderSupport is called by Property loader spring bean.