number of nodes limit a cluster
The recommendation to use less then 8 nodes in a Hazelcast cluster was based on the number of nodes that had been tested by QA. There is not real evidence for this limit. Any limitation is a by-product of Hazelcast itself and the message size. It is not dependent on MuleSoft's implementation or use of Hazelcast.