copyedit and temporarily hiding content until...
I'd like to address the TODO I added to the hidden content. Essentially, I'd like to MOVE MOST OF THIS TO ITS OWN TOPIC AND RESOLVE OVERLAP WITH OTHER CONTENT TO AVOID DUPLICATION. SOME OF IT TALKS ABOUT STUFF NOT YET INTRODUCED (readers/writers). I'll be in contact with Brett to determine the appropriate level of detail here vs. the new topic.