Remove duplicated instructions
There were two sections explaining pretty much the same thing. I've removed one of them, and added a small comment beforehand to point to it, and let the user know that they will still need to manually classify the API after the runtime has been configured with the environment credentials.